Docker initial setup
Should work, but mounting the source to the image is ... a workaround. Use it as a copy, and to keep dbconnect.php persistent, as well as any modules you put in.
Not elegant, but else we'd need a conf-dir, which would make bigger changes and we lose any compatibility with old versions.

Also several bugfixes, especially on the installer that could hang in one instance, again.